pyproject.toml
[project]
name = "movfx"
version = "0.1.0"  # semantic-release
description = "MovFX provides transition effects between images with optional sound."
readme = "docs/index.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10,<4"
authors = [
  {name = "Ivan Ogasawara", email = "ivan.ogasawara@gmail.com"},
]
dependencies = [
  "moviepy>=2.0",
  "pillow>=10.0",
  "numpy>=1.20,<3",
  "typer>=0.9.0",
  "pyyaml>=6.0",
]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
  {include = "movfx", from = "src"},
]
exclude = [
  ".git/*",
  ".env*",
]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.2"
pytest-cov = ">=4.1.0"
coverage = ">=7.2.7"
pre-commit = ">=3.3.2"
ruff = ">=0.1.5"
mypy = ">=1.6.0"
bandit = ">=1.7.5"
vulture = ">=2.9.1"
mccabe = ">=0.6.1"
ipykernel = ">=6.0.0"
ipython = ">=7.0.0"
mkdocs = ">=1.3"
mkdocs-exclude = ">=1.0.2"
mkdocs-jupyter = ">=0.24.7"
mkdocs-literate-nav = ">=0.4.1"
mkdocs-macros-plugin = ">=0.6.3"
mkdocs-material = ">=8"
mkdocstrings = {version=">=0.19.0", extras=["python"]}
mkdocstrings-python = "<2"
jupyterlab = ">=4.0.5"
makim = "1.13.0"
douki = ">=0.6.0"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.vulture]
exclude = ["tests"]
ignore_decorators = []
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["src/movfx"]
sort_by_size = true
verbose = false

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
exclude = [
  'docs',
]
fix = true

[tool.ruff.lint]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  # "D",   # pydocstyle (disabled: using douki YAML docstrings)
  "YTT", # flake8-2020
  "PL",  # PL
  "RUF", # Ruff-specific rules
  "I001", # isort
]
ignore = ["PLR0913", "PLR0915"]

# [tool.ruff.lint.pydocstyle]  # disabled: using douki YAML docstrings
# convention = "numpy"

[tool.ruff.lint.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.ruff.format]
quote-style = "double"


[tool.mypy]
no_strict_optional = false

[[tool.mypy.overrides]]
module = [
  "moviepy",
  "moviepy.*",
  "pydub",
  "pydub.*",
  "yaml",
]
ignore_missing_imports = true
